Reorganized with atomic design.
