+++ Job-listings-app/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './components/App.jsx'
+import App from './views/App.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
